Refactor to prepare for more api's
